Wow my first platform specific fix

+++ view/layout.go
@@ -87,7 +87,7 @@ func statusLine() string {
 		log.Print(err)
 	}
 
-	freeStorage := float64(fsStat.Bavail*uint64(fsStat.Bsize)) / float64(1024*1024*1024)
+	freeStorage := float64(uint64(fsStat.Bavail)*uint64(fsStat.Bsize)) / float64(1024*1024*1024)
 	memoryUsed := float64(memStat.Alloc) / float64(1024*1024)
 
 	return fmt.Sprintf("OS:%v/%v FreeStorage:%.2f Gb, MemUsed: %.2f Mb", runtime.GOOS, runtime.GOARCH, freeStorage, memoryUsed)
